apps/ai-service/pyproject.toml
[project]
name = "ai-service"
version = "0.1.0"
description = "FastAPI ML Recommendation Service"
authors = [
    {name = "Admin", email = "admin@example.com"}
]
requires-python = ">=3.11"
dependencies = [
    "fastapi==0.111.0",
    "uvicorn[standard]>=0.30.0",
    "pydantic-settings>=2.2.1",
    "lightfm-next",
    "scikit-learn==1.4.0",
    "numpy>=1.26.4",
    "scipy>=1.13.0",
    "pymongo>=4.6.3",
    "motor>=3.4.0",
    "boto3>=1.34.0",
    "httpx>=0.27.0"
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"
